fix(count): count models added
