+++ app/src/main/java/com/polina/fintrackr/features/incoms/domain/Count.kt
@@ -1,0 +1,11 @@
+package com.polina.fintrackr.features.incoms.domain
+
+data class Count(
+    val id: Int = 1,
+    val userId: Int = 1,
+    val name: String = "",
+    val balance: String = "",
+    val currency: String = "",
+    val createdAt: String = "",
+    val updatedAt: String = ""
+)
